% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{from_json}
\alias{from_json}
\title{Parse JSON String}
\usage{
from_json(json)
}
\arguments{
\item{json}{const char, JSON string to be parsed. Coming from R, this
input should be a character vector of length 1.}
}
\description{
Takes a JSON string as input, returns a list of key-value pairs
}
